feat(orders): accept decimal quantities for weighable units in wizard
OrderCreationWizard hard-coded `parseInt` + `min={1}` for both the
picker quantity input and the per-row cart input, so products sold in
kg, L, m or m² could not be ordered in the fractional amounts the
backend already accepts (and `validateQuantityForUnit` validates).

Now the wizard reads `product.unit`:
- Picker input shows "Qtd (kg)" / "Qtd (L)" when the selected product
  is weighable, with `step` = 1 / 10**decimals and `min` = step. Falls
  back to integer Qtd for un/cx/pc.
- Per-row cart input uses the same logic against the row's product
  unit, so a cart can mix integer and decimal items consistently.
- `handleUpdateQuantity` floor is the unit step instead of a hard 1,
  so 0.250 kg no longer reverts to 1.

Layout: weighable rows widen the input from w-12 to w-16 so 3 decimal
digits + a "." fit without clipping.

`unitAbbrev` / `unitDecimals` / `isDecimalUnit` imported from the
canonical registry, matching the POS scale fix and backend validation.

+++ backend/src/validation/fiscal.ts
@@ -44,8 +44,8 @@ export const createMedicationSchema = z.object({
     expiryDate: z.string().datetime({ message: 'Data de validade inválida' }),
     price: z.number().positive('Preço deve ser maior que zero'),
     costPrice: z.number().min(0, 'Custo não pode ser negativo').optional().default(0),
-    stock: z.number().int().min(0, 'Stock não pode ser negativo').optional().default(0),
-    minStock: z.number().int().min(0, 'Stock mínimo não pode ser negativo').optional().default(5),
+    stock: z.number().min(0, 'Stock não pode ser negativo').optional().default(0),
+    minStock: z.number().min(0, 'Stock mínimo não pode ser negativo').optional().default(5),
     requiresPrescription: z.boolean().optional().default(false),
     controlled: z.boolean().optional().default(false),
     storageConditions: z.string().max(200, 'Condições de armazenamento muito longas').optional().nullable(),
@@ -64,7 +64,7 @@ export const createPrescriptionSchema = z.object({
     prescriptionDate: z.string().datetime({ message: 'Data da prescrição inválida' }),
     items: z.array(z.object({
         medicationId: z.string().uuid('ID do medicamento inválido'),
-        quantity: z.number().int().positive('Quantidade deve ser maior que zero'),
+        quantity: z.number().positive('Quantidade deve ser maior que zero'),
         dosage: z.string().max(200, 'Dosagem muito longa').optional().nullable(),
         frequency: z.string().max(100, 'Frequência muito longa').optional().nullable(),
         duration: z.string().max(100, 'Duração muito longa').optional().nullable(),
@@ -75,7 +75,7 @@ export const createPrescriptionSchema = z.object({
 
 export const dispenseMedicationSchema = z.object({
     medicationId: z.string().uuid('ID do medicamento inválido'),
-    quantity: z.number().int().positive('Quantidade deve ser maior que zero'),
+    quantity: z.number().positive('Quantidade deve ser maior que zero'),
     patientName: z.string().max(200, 'Nome do paciente muito longo').optional().nullable(),
     prescriptionId: z.string().uuid('ID da prescrição inválido').optional().nullable(),
     notes: z.string().max(500, 'Notas muito longas').optional().nullable()
